slack-bot/requirements.txt
# Core Slack SDK
slack-bolt>=1.18.0
slack-sdk>=3.26.0
aiohttp>=3.9.0

# SSL certificates
certifi>=2023.0.0

# Redis for distributed rate limiting (optional)
redis>=5.0.0

# Environment management
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.7

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
